+++ rye/src/installer.rs
@@ -1,5 +1,6 @@
+use std::collections::BTreeMap;
 use std::os::unix::fs::symlink;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::process::{Command, Stdio};
 use std::{env, fs};
 
@@ -19,12 +20,42 @@ use crate::utils::CommandOutput;
 
 const FIND_SCRIPT_SCRIPT: &str = r#"
 import os
+import re
 import sys
-from importlib.metadata import distribution
+import json
+from importlib.metadata import distribution, PackageNotFoundError
 
-dist = distribution(sys.argv[1])
-for file in dist.files:
-    print(os.path.normpath(dist.locate_file(file)))
+_package_re = re.compile('(?i)^([a-z0-9._-]+)')
+
+result = {}
+
+def dump_all(dist, root=False):
+    rv = []
+    for file in dist.files or ():
+        rv.append(os.path.normpath(dist.locate_file(file)))
+    result["" if root else dist.name] = rv
+    req = []
+    for r in dist.requires or ():
+        name = _package_re.match(r)
+        if name is not None:
+            req.append(name.group())
+    return req
+
+root = sys.argv[1]
+to_resolve = [root]
+seen = set()
+while to_resolve:
+    try:
+        d = to_resolve.pop()
+        dist = distribution(d)
+    except Exception:
+        continue
+    if dist.name in seen:
+        continue
+    seen.add(dist.name)
+    to_resolve.extend(dump_all(dist, root=d==root))
+
+print(json.dumps(result))
 "#;
 static SUCCESSFULLY_DOWNLOADED_RE: Lazy<Regex> =
     Lazy::new(|| Regex::new("(?m)^Successfully downloaded (.*?)$").unwrap());
@@ -33,12 +64,17 @@ pub fn install(
     requirement: Requirement,
     py_ver: &PythonVersionRequest,
     force: bool,
+    include_deps: &[String],
     output: CommandOutput,
 ) -> Result<(), Error> {
     let app_dir = get_app_dir()?;
     let shim_dir = app_dir.join("shims");
     let self_venv = ensure_self_venv(output)?;
     let tool_dir = app_dir.join("tools");
+    let include_deps = include_deps
+        .iter()
+        .map(|x| normalize_package_name(x))
+        .collect::<Vec<_>>();
 
     let target_venv_path = tool_dir.join(normalize_package_name(&requirement.name));
     if target_venv_path.is_dir() && !force {
@@ -53,7 +89,7 @@ pub fn install(
 
     create_virtualenv(output, &self_venv, &py_ver, &target_venv_path)?;
 
-    let mut cmd = Command::new(&self_venv.join("bin/pip"));
+    let mut cmd = Command::new(self_venv.join("bin/pip"));
     cmd.arg("--python")
         .arg(&target_venv_bin_path.join("python"))
         .arg("install")
@@ -73,40 +109,110 @@ pub fn install(
         bail!("tool installation failed");
     }
 
-    let out = Command::new(&target_venv_bin_path.join("python"))
+    let out = Command::new(target_venv_bin_path.join("python"))
         .arg("-c")
         .arg(FIND_SCRIPT_SCRIPT)
         .arg(&requirement.name)
         .stdout(Stdio::piped())
         .output()
         .context("unable to dump package manifest from installed package")?;
-    let files = std::str::from_utf8(&out.stdout)
-        .context("non utf-8 package manifest")?
-        .lines()
-        .map(Path::new)
-        .collect::<Vec<_>>();
+    let all_files: BTreeMap<String, Vec<PathBuf>> = serde_json::from_slice(&out.stdout)
+        .with_context(|| {
+            format!(
+                "failed to resolve manifest\n{}",
+                String::from_utf8_lossy(&out.stderr)
+            )
+        })?;
 
-    let mut installed_any = false;
-    for file in files {
-        if let Ok(rest) = file.strip_prefix(&target_venv_bin_path) {
-            let shim_target = shim_dir.join(rest);
-            symlink(file, shim_target)
-                .with_context(|| format!("unable to symlink tool to {}", file.display()))?;
-            installed_any = true;
-            if output != CommandOutput::Quiet {
-                eprintln!("installed script {}", style(rest.display()).cyan());
+    let mut installed = Vec::new();
+    let mut scripts_found = Vec::new();
+    if let Some(files) = all_files.get("") {
+        installed.extend(install_scripts(files, &target_venv_bin_path, &shim_dir)?);
+    }
+
+    for (package, files) in all_files.iter() {
+        if package.is_empty() {
+            continue;
+        }
+        if include_deps.contains(&normalize_package_name(package)) {
+            installed.extend(install_scripts(files, &target_venv_bin_path, &shim_dir)?);
+        } else {
+            let scripts = find_scripts(files, &target_venv_bin_path);
+            if !scripts.is_empty() {
+                scripts_found.push((package, scripts));
             }
         }
     }
 
-    if !installed_any && output != CommandOutput::Quiet {
+    if !scripts_found.is_empty()
+        && output != CommandOutput::Quiet
+        && (installed.is_empty() || output == CommandOutput::Verbose)
+    {
         eprintln!(
             "{}",
-            style("warning: installed package did not expose any scripts").red()
+            style("Found additional non installed scripts in dependencies:").yellow()
         );
+        scripts_found.sort();
+        for (package, scripts) in scripts_found.iter() {
+            eprintln!("{}:", style(package).green());
+            for script in scripts {
+                eprintln!("  - {}", style(script).cyan());
+            }
+        }
+        eprintln!("To install scripts from these packages pass the appropriate --include-dep");
+    }
+
+    if output != CommandOutput::Quiet {
+        eprintln!();
+        if installed.is_empty() {
+            eprintln!(
+                "{}",
+                style("warning: installed package did not expose any scripts").red()
+            );
+        } else {
+            eprintln!("Installed scripts:");
+            for script in installed {
+                eprintln!("  - {}", style(script).cyan());
+            }
+            if output != CommandOutput::Verbose && !scripts_found.is_empty() {
+                eprintln!();
+                eprintln!(
+                    "note: {}",
+                    style("additional scripts were encountered in non-installed dependencies.")
+                        .dim()
+                );
+            }
+        }
     }
 
     Ok(())
+}
+
+fn find_scripts(files: &[PathBuf], target_venv_bin_path: &Path) -> Vec<String> {
+    let mut rv = Vec::new();
+    for file in files {
+        if let Ok(rest) = file.strip_prefix(target_venv_bin_path) {
+            rv.push(rest.to_string_lossy().to_string());
+        }
+    }
+    rv
+}
+
+fn install_scripts(
+    files: &[PathBuf],
+    target_venv_bin_path: &Path,
+    shim_dir: &Path,
+) -> Result<Vec<String>, Error> {
+    let mut rv = Vec::new();
+    for file in files {
+        if let Ok(rest) = file.strip_prefix(target_venv_bin_path) {
+            let shim_target = shim_dir.join(rest);
+            symlink(file, shim_target)
+                .with_context(|| format!("unable to symlink tool to {}", file.display()))?;
+            rv.push(rest.to_string_lossy().to_string());
+        }
+    }
+    Ok(rv)
 }
 
 pub fn uninstall(package: &str, output: CommandOutput) -> Result<(), Error> {